feat: Move AppUpdate to kmpuiviews and Remove SettingsViewModel
This commit moves the `AppUpdate` functionality to the `kmpuiviews` module for multiplatform use, removes the Android-specific `SettingsViewModel`, and updates related dependencies.

-   **AppUpdate:**
    -   Moved `AppUpdate` from `sharedutils` to `kmpuiviews`.
    -   Removed the old `AppUpdate` from `sharedutils`.
    -   Added `AppUpdateCheck` to `kmpuiviews`.
    - Updated how the app checks for an update.
-    - Updated the dependencies to the new `AppUpdate`.
+- **SettingsViewModel:**
+    - Removed `SettingsViewModel` from `UIViews`.
+- **Other**
+   - Updated the imports related to `AppUpdate`.
+   - Updated `AppCheckWorker` to use the new `AppUpdate`.
+   - Updated `InfoSettings` to use the new `AppUpdate`.
+   - Updated `MoreInfoViewModel` to use the new `AppUpdateCheck`.
+   - Updated the DI modules to reflect these changes.
+   - General cleanup.
