+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/InfoSettings.kt
@@ -36,6 +36,8 @@ import com.google.accompanist.drawablepainter.rememberDrawablePainter
 import com.programmersbox.datastore.DataStoreHandling
 import com.programmersbox.datastore.asState
 import com.programmersbox.helpfulutils.requestPermissions
+import com.programmersbox.kmpuiviews.domain.AppUpdate
+import com.programmersbox.kmpuiviews.domain.AppUpdateCheck
 import com.programmersbox.kmpuiviews.platform
 import com.programmersbox.kmpuiviews.presentation.Screen
 import com.programmersbox.kmpuiviews.presentation.components.PreferenceSetting
@@ -45,8 +47,6 @@ import com.programmersbox.kmpuiviews.utils.LocalNavController
 import com.programmersbox.kmpuiviews.utils.composables.icons.Discord
 import com.programmersbox.kmpuiviews.utils.composables.icons.Github
 import com.programmersbox.sharedutils.AppLogo
-import com.programmersbox.sharedutils.AppUpdate
-import com.programmersbox.sharedutils.updateAppCheck
 import com.programmersbox.uiviews.BuildConfig
 import com.programmersbox.uiviews.R
 import com.programmersbox.uiviews.presentation.settings.viewmodels.MoreInfoViewModel
@@ -74,6 +74,7 @@ fun InfoSettings(
     val context = LocalContext.current
     val uriHandler = LocalUriHandler.current
     val dataStoreHandling = koinInject<DataStoreHandling>()
+    val appUpdateCheck: AppUpdateCheck = koinInject()
 
     SettingsScaffold(stringResource(R.string.more_info_category)) {
         PreferenceSetting(
@@ -144,7 +145,7 @@ fun InfoSettings(
             ) { uriHandler.openUri("https://ko-fi.com/V7V3D3JI") }
         )
 
-        val appUpdate by updateAppCheck.collectAsStateWithLifecycle(null)
+        val appUpdate by appUpdateCheck.updateAppCheck.collectAsStateWithLifecycle(null)
 
         PreferenceSetting(
             settingIcon = {
@@ -188,7 +189,9 @@ fun InfoSettings(
                                     Manifest.permission.READ_EXTERNAL_STORAGE
                                 ) {
                                     if (it.isGranted) {
-                                        updateAppCheck.value
+                                        appUpdateCheck
+                                            .updateAppCheck
+                                            .value
                                             ?.let { a ->
                                                 infoViewModel.update(a)
                                             }
