+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/viewmodels/MoreInfoViewModel.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import android.widget.Toast
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.programmersbox.sharedutils.AppUpdate
-import com.programmersbox.sharedutils.updateAppCheck
+import com.programmersbox.kmpuiviews.domain.AppUpdate
+import com.programmersbox.kmpuiviews.domain.AppUpdateCheck
 import com.programmersbox.uiviews.GenericInfo
 import com.programmersbox.uiviews.presentation.settings.downloadstate.DownloadAndInstaller
 import kotlinx.coroutines.Dispatchers
@@ -17,6 +17,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 class MoreInfoViewModel(
     val downloadAndInstaller: DownloadAndInstaller,
     private val genericInfo: GenericInfo,
+    private val appUpdateCheck: AppUpdateCheck,
 ) : ViewModel() {
 
     private val checker = AtomicBoolean(false)
@@ -40,7 +41,7 @@ class MoreInfoViewModel(
         try {
             if (!checker.get()) {
                 checker.set(true)
-                AppUpdate.getUpdate()?.let(updateAppCheck::tryEmit)
+                AppUpdate.getUpdate()?.let(appUpdateCheck.updateAppCheck::tryEmit)
             }
         } catch (e: Exception) {
             e.printStackTrace()
